+++ protocols/gossipsub/CHANGELOG.md
@@ -11,6 +11,12 @@
 
 [PR 4648]: (https://github.com/libp2p/rust-libp2p/pull/4648)
 
+<!-- Internal changes
+
+- Allow deprecated usage of `KeepAlive::Until`
+
+-->
+
 ## 0.45.1
 
 - Add getter function to obtain `TopicScoreParams`.